Move common resources to shared folder
