+++ DX11Renderer/shared/CheckHr.h
