+++ DX11Renderer/shared/MeshVertexShader.hlsl
