+++ DX11Renderer/shared/NormalPixelShader.hlsl
