+++ DX11Renderer/shared/NormalVertexShader.hlsl
