+++ DX11Renderer/shared/WinApp.cpp
